tools/pipeline_runner/pyproject.toml
[tool.poetry]
name = "pipeline-runner"
version = "0.1.0"
description = "Generic pipeline runner with step-based execution, YAML definitions, and structured observability."
authors = ["OpenClaw Contributors"]
readme = "README.md"
packages = [{include = "pipeline_runner"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"

[tool.poetry.scripts]
pipeline-runner = "pipeline_runner.runner:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
